# apps/gitpaywidget/openapi.yaml
openapi: 3.1.0
info:
  title: GitPayWidget API
  description: |
    REST API for GitPayWidget - the plug-and-play payment solution for static sites.
    
    ## Authentication
    Most endpoints require Supabase session authentication via cookies. 
    The SDK handles this automatically when users sign in with GitHub OAuth.
    
    ## Rate Limits
    - Public endpoints: 60 req/min per IP
    - Authenticated endpoints: 120 req/min per user
    - Checkout endpoint: 10 req/min per IP
    - Webhook endpoint: No limit (verified via signature)
    
    ## Base URL
    Production: `https://gitpaywidget.com/api`
  version: 0.1.0
  contact:
    name: Manic Agency
    email: team@manic.agency
    url: https://manic.agency
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://gitpaywidget.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Local development
...
paths:
  /checkout:
    post:
      tags: [Checkout]
      summary: Create checkout session
      description: |
        Create a hosted checkout session with Stripe or Lemon Squeezy.
        Returns a URL to redirect the customer to the payment page.
      operationId: createCheckout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
      responses:
        '200':
          description: Checkout session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutResponse'
        '400':
          description: Invalid request or missing provider credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'
        '500':
          description: Provider API error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /webhook:
    post:
      tags: [Webhooks]
      summary: Receive webhook events
      description: |
        Receive webhook events from Stripe or Lemon Squeezy.
        Verifies signature and processes the event.
      operationId: handleWebhook
      parameters:
        - name: X-GPW-Provider
          in: header
          schema:
            type: string
            enum: [stripe, lemonsqueezy]
          description: Provider identifier (auto-detected if omitted)
        - name: Stripe-Signature
          in: header
          schema:
            type: string
          description: Stripe webhook signature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Event processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
        '400':
          description: Invalid signature or payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase session cookie
  schemas:
    CheckoutRequest:
      type: object
      required: [project, plan]
      properties:
        project:
          type: string
          description: Project slug (e.g., "acme/site")
          example: "myorg/landing"
        plan:
          type: string
          description: Plan identifier
          example: "pro"
        metadata:
          type: object
          additionalProperties:
            type: string
          description: Additional metadata for the checkout
          example:
            userId: "12345"
            referrer: "landing-page"
    CheckoutResponse:
      type: object
      properties:
        checkoutUrl:
          type: string
          format: uri
          description: URL to redirect customer to
        sessionId:
          type: string
          description: Checkout session ID
        provider:
          type: string
          enum: [stripe, lemonsqueezy]
          description: Payment provider used
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
    RateLimitError:
      type: object
      properties:
        error:
          type: string
          example: "Too many requests"
        retryAfter:
          type: integer
          description: Seconds until rate limit resets